Initial check in of the first incarnation of the source.
This source leaves a lot to be desired and will no doubt change alot.
Currently there is a basic TODO in the source, this will be moved out
to a separate file, but in the meantime the first things it summarizes
are adding ID3v1 extended support, write capacity, making this a gem, and
also cleaning up the code a bit. Once that is achieved, then it will be
time to reassess and move foreward.